# .docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/catalog/items": {
      "post": {
        "operationId": "CatalogController_registerItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCatalogItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item created or updated"
          }
        },
        "summary": "Create or update compliance item",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/catalog/merchants/{merchantId}/items": {
      "get": {
        "operationId": "CatalogController_listItems",
        "parameters": [
          {
            "name": "merchantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item list"
          }
        },
        "summary": "List compliance items",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/catalog/items/sync": {
      "post": {
        "operationId": "CatalogController_syncItems",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncCatalogItemsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Items synced"
          }
        },
        "summary": "Sync catalog items to eTIMS (register/update)",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/catalog/item-classifications": {
      "get": {
        "operationId": "CatalogController_searchItemClassifications",
        "parameters": [
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Filter by item classification code prefix or name (case-insensitive contains)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemClsLvl",
            "required": false,
            "in": "query",
            "description": "Filter by classification level (1-5)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "description": "Include codes where useYn=N (retired). Defaults to false.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max rows to return (default 20, max 100)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching classification codes"
          }
        },
        "summary": "Search OSCU item classification codes (itemClsCd) for use when registering items",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/catalog/item-classifications/{itemClsCd}": {
      "get": {
        "operationId": "CatalogController_getItemClassification",
        "parameters": [
          {
            "name": "itemClsCd",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Classification code"
          }
        },
        "summary": "Fetch a single OSCU item classification code",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/catalog/item-classifications/sync": {
      "post": {
        "operationId": "CatalogController_syncItemClassifications",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncItemClassificationsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sync result"
          }
        },
        "summary": "Pull the latest item classification reference list from OSCU (/selectItemClsList) and upsert it locally",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/catalog/code-classes": {
      "get": {
        "operationId": "CatalogController_listCodeClasses",
        "parameters": [
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "description": "Include groups where useYn=N (retired). Defaults to false.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Code groups"
          }
        },
        "summary": "List OSCU code groups (cdCls), e.g. '10' Unit of Quantity, '17' Packaging Unit, '04' Tax Type",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/catalog/codes": {
      "get": {
        "operationId": "CatalogController_searchCodes",
        "parameters": [
          {
            "name": "cdCls",
            "required": false,
            "in": "query",
            "description": "Code group to search within, e.g. '10' (Unit of Quantity), '17' (Packaging Unit), '04' (Tax Type), '24' (Product Type), '07' (Payment Method). Omit to search across all groups.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Filter by code prefix or name (case-insensitive contains)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "description": "Include codes where useYn=N (retired). Defaults to false.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max rows to return (default 50, max 200)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching codes"
          }
        },
        "summary": "Search OSCU codes (qtyUnitCd, pkgUnitCd, taxTyCd, itemTyCd, pmtTyCd...) within a code group",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/catalog/codes/sync": {
      "post": {
        "operationId": "CatalogController_syncCodeList",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncCodeListDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sync result"
          }
        },
        "summary": "Pull the latest code list from OSCU (/selectCodeList) and upsert it locally",
        "tags": [
          "Catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sync2Books Compliance API",
    "description": "Unified compliance abstraction layer for eTIMS (OSCU).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "RegisterCatalogItemDto": {
        "type": "object",
        "properties": {
          "merchantId": {
            "type": "string"
          },
          "externalId": {
            "type": "string",
            "description": "ERP/external item id (e.g. QuickBooks Item.Id)"
          },
          "name": {
            "type": "string"
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "itemType": {
            "type": "string",
            "enum": [
              "GOODS",
              "SERVICE"
            ]
          },
          "taxCategory": {
            "type": "string",
            "enum": [
              "VAT_STANDARD",
              "VAT_ZERO",
              "EXEMPT",
              "OTHER"
            ]
          },
          "classificationCode": {
            "type": "string",
            "description": "OSCU item classification code (itemClsCd) override"
          },
          "unitCode": {
            "type": "string",
            "description": "OSCU unit of quantity code (qtyUnitCd) override"
          },
          "internalUnit": {
            "type": "string",
            "description": "Internal/ERP unit identifier (EA, PCS, EACH...)"
          },
          "packagingUnitCode": {
            "type": "string",
            "description": "OSCU packaging unit code (pkgUnitCd) override"
          },
          "taxTyCd": {
            "type": "string",
            "description": "OSCU tax type code (taxTyCd) override"
          },
          "productTypeCode": {
            "type": "string",
            "description": "OSCU product type code (itemTyCd) override"
          }
        },
        "required": [
          "merchantId",
          "externalId",
          "name",
          "itemType",
          "taxCategory"
        ]
      },
      "SyncCatalogItemsDto": {
        "type": "object",
        "properties": {
          "merchantId": {
            "type": "string"
          },
          "branchId": {
            "type": "string"
          },
          "itemIds": {
            "description": "Optional list of internal item ids to sync.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "onlyPending": {
            "type": "boolean",
            "default": true,
            "description": "When true, only sync items in PENDING/FAILED state (skips REGISTERED)."
          },
          "force": {
            "type": "boolean",
            "default": false,
            "description": "When true, force re-sync even if already REGISTERED."
          }
        },
        "required": [
          "merchantId",
          "branchId"
        ]
      },
      "SyncItemClassificationsDto": {
        "type": "object",
        "properties": {
          "merchantId": {
            "type": "string",
            "description": "Merchant with a provisioned eTIMS connection to use for the OSCU call. The fetched list is global reference data, not merchant-scoped."
          },
          "branchId": {
            "type": "string"
          },
          "full": {
            "type": "boolean",
            "description": "Ignore the stored watermark (lastReqDt) and re-pull the full reference list."
          }
        },
        "required": [
          "merchantId",
          "branchId"
        ]
      },
      "SyncCodeListDto": {
        "type": "object",
        "properties": {
          "merchantId": {
            "type": "string",
            "description": "Merchant with a provisioned eTIMS connection to use for the OSCU call. The fetched list is global reference data, not merchant-scoped."
          },
          "branchId": {
            "type": "string"
          },
          "full": {
            "type": "boolean",
            "description": "Ignore the stored watermark (lastReqDt) and re-pull the full reference list."
          }
        },
        "required": [
          "merchantId",
          "branchId"
        ]
      },
...
    }
  }
}